#1693 move getDateFromTimeString to dateTimeUtils
