+++ wls-gui-wahllokalsystem/src/components/common/inputs/BaseTimeInput.vue
@@ -14,11 +14,13 @@
 import type { PropType } from "vue";
 
 import { useDateTimeFormatter } from "@/composables/common/dateTimeFormatter.ts";
+import { useDateTimeUtils } from "@/composables/common/dateTimeUtils.ts";
 import { useRules } from "@/composables/common/rules.ts";
 
 const { required } = useRules();
 
-const { toHhMm, getDateFromTimeString } = useDateTimeFormatter();
+const { toHhMm } = useDateTimeFormatter();
+const { getDateFromTimeString } = useDateTimeUtils();
 
 const modelValue = defineModel({
   type: Object as PropType<Date>,
