+++ wls-gui-wahllokalsystem/src/composables/common/dateTimeUtils.ts
@@ -12,8 +12,39 @@ export function useDateTimeUtils() {
     return mappedUhrzeit;
   };
 
+  const getDateFromTimeString = function (timeString: string): Date {
+    // Validate time string format (HH:MM)
+    if (!timeString || !/^\d{1,2}:\d{1,2}(?::\d{1,2})?$/.test(timeString)) {
+      return new Date(NaN);
+    }
+
+    const timeParts = timeString.split(":").map(Number);
+
+    const [hours, minutes] = timeParts;
+    if (
+      isNaN(hours) ||
+      isNaN(minutes) ||
+      hours < 0 ||
+      hours > 23 ||
+      minutes < 0 ||
+      minutes > 59
+    ) {
+      return new Date(NaN); // Return invalid date for invalid time values
+    }
+
+    const seconds = timeParts.length === 3 ? timeParts[2] : 0;
+    if (!isNaN(seconds) && (seconds < 0 || seconds > 59)) {
+      return new Date(NaN);
+    }
+
+    const date = new Date();
+    date.setHours(hours, minutes, seconds, 0);
+    return date;
+  };
+
   return {
     isValidDate,
     applyLocalTimezoneOffset,
+    getDateFromTimeString,
   };
 }
