+++ wls-gui-wahllokalsystem/tests/composables/common/dateTimeUtils.spec.ts
@@ -2,7 +2,8 @@ import { describe, expect, it } from "vitest";
 
 import { useDateTimeUtils } from "@/composables/common/dateTimeUtils.ts";
 
-const { isValidDate, applyLocalTimezoneOffset } = useDateTimeUtils();
+const { isValidDate, applyLocalTimezoneOffset, getDateFromTimeString } =
+  useDateTimeUtils();
 
 describe("dateTimeUtils", () => {
   describe("isValidDate", () => {
@@ -41,5 +42,35 @@ describe("dateTimeUtils", () => {
       expect(result).toBeInstanceOf(Date);
       expect(isNaN(result.getTime())).toBe(true);
     });
+  });
+
+  describe("getDateFromTimeString", () => {
+    it.each(["12:12", "12:12:30"])(
+      "should_returnDateWithGivenTime_when_givenValidTimeString'%s'",
+      (input) => {
+        const result = getDateFromTimeString(input).toString();
+
+        expect(result).toContain(input);
+      }
+    );
+
+    it.each(["text", "26:12", "11:78", "23:45:67", "", " "])(
+      "should_returnInvalidDate_when_inputStringIsInvalidValue'%s'",
+      (input) => {
+        const result = getDateFromTimeString(input);
+
+        expect(result).toBeInstanceOf(Date);
+        expect(isNaN(result.getTime())).toBe(true);
+      }
+    );
+
+    it.each(["12:12", "12:12:30"])(
+      "should_returnDateWithZeroMilliseconds_when_givenValidTimeString'%s'",
+      (input) => {
+        const result = getDateFromTimeString(input);
+
+        expect(result.getMilliseconds()).toBe(0);
+      }
+    );
   });
 });
